requirements.txt
asgiref==3.4.1
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.3
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
typing-extensions==3.10.0.0